+++ api/src/__init__.py
@@ -1,194 +1,117 @@
-import uuid
-from datetime import datetime
-from typing import Any, Dict, Iterator, List, Optional
-
-from api.src import Comment, Issue, IssueTrackerClient
+from typing import Any, Dict, Iterator, List, Optional, Protocol, runtime_checkable
 
 
-class MemoryComment(Comment):
-    """An in-memory implementation of a Comment."""
-    
-    def __init__(self, author: str, content: str):
-        self._id = str(uuid.uuid4())
-        self._author = author
-        self._content = content
-        self._created_at = datetime.now().isoformat()
-    
+@runtime_checkable
+class Comment(Protocol):
+    """A comment on an issue."""
+
     @property
     def id(self) -> str:
-        return self._id
-    
+        """Return the id of the comment."""
+        raise NotImplementedError
+
     @property
     def author(self) -> str:
-        return self._author
-    
+        """Return the author of the comment."""
+        raise NotImplementedError
+
     @property
     def content(self) -> str:
-        return self._content
-    
+        """Return the content of the comment."""
+        raise NotImplementedError
+
     @property
     def created_at(self) -> str:
-        return self._created_at
+        """Return the creation date of the comment."""
+        raise NotImplementedError
 
+@runtime_checkable
+class Issue(Protocol):
+    """An Issue in the issue tracker."""
 
-class MemoryIssue(Issue):
-    """An in-memory implementation of an Issue."""
-    
-    def __init__(self, title: str, description: str, creator: str, **kwargs):
-        self._id = str(uuid.uuid4())
-        self._title = title
-        self._description = description
-        self._status = kwargs.get("status", "open")
-        self._creator = creator
-        self._assignee = kwargs.get("assignee")
-        self._created_at = datetime.now().isoformat()
-        self._updated_at = self._created_at
-        self._labels = kwargs.get("labels", [])
-        self._priority = kwargs.get("priority")
-        self._comments: List[MemoryComment] = []
-    
     @property
     def id(self) -> str:
-        return self._id
-    
+        """Return the id of the issue."""
+        raise NotImplementedError
+
     @property
     def title(self) -> str:
-        return self._title
-    
+        """Return the title of the issue."""
+        raise NotImplementedError
+
     @property
     def description(self) -> str:
-        return self._description
-    
+        """Return the description of the issue."""
+        raise NotImplementedError
+
     @property
     def status(self) -> str:
-        return self._status
-    
+        """Return the status of the issue (open, closed, etc.)."""
+        raise NotImplementedError
+
     @property
     def creator(self) -> str:
-        return self._creator
-    
+        """Return the creator of the issue."""
+        raise NotImplementedError
+
     @property
     def assignee(self) -> Optional[str]:
-        return self._assignee
-    
+        """Return the assignee of the issue, if any."""
+        raise NotImplementedError
+
     @property
     def created_at(self) -> str:
-        return self._created_at
-    
+        """Return the creation date of the issue."""
+        raise NotImplementedError
+
     @property
     def updated_at(self) -> str:
-        return self._updated_at
-    
+        """Return the last update date of the issue."""
+        raise NotImplementedError
+
     @property
     def labels(self) -> List[str]:
-        return self._labels
-    
+        """Return the labels associated with the issue."""
+        raise NotImplementedError
+
     @property
     def priority(self) -> Optional[str]:
-        return self._priority
-    
+        """Return the priority of the issue, if set."""
+        raise NotImplementedError
+
     def get_comments(self) -> Iterator[Comment]:
-        return iter(self._comments)
-    
-    def add_comment(self, comment: MemoryComment) -> None:
-        self._comments.append(comment)
-        self._updated_at = datetime.now().isoformat()
-    
-    def update(self, **kwargs) -> None:
-        """Update issue attributes."""
-        if "title" in kwargs:
-            self._title = kwargs["title"]
-        if "description" in kwargs:
-            self._description = kwargs["description"]
-        if "status" in kwargs:
-            self._status = kwargs["status"]
-        if "assignee" in kwargs:
-            self._assignee = kwargs["assignee"]
-        if "labels" in kwargs:
-            self._labels = kwargs["labels"]
-        if "priority" in kwargs:
-            self._priority = kwargs["priority"]
-        
-        self._updated_at = datetime.now().isoformat()
+        """Return an iterator of comments for this issue."""
+        raise NotImplementedError
 
+@runtime_checkable
+class IssueTrackerClient(Protocol):
+    """An Issue Tracker Client used to manage issues."""
 
-class MemoryIssueTrackerClient(IssueTrackerClient):
-    """An in-memory implementation of an Issue Tracker Client."""
-    
-    def __init__(self):
-        self._issues: Dict[str, MemoryIssue] = {}
-        self._current_user = "default_user"  # In a real system, this would come from auth
-    
-    def set_current_user(self, username: str) -> None:
-        """Set the current user for operations."""
-        self._current_user = username
-    
     def get_issues(self, filters: Optional[Dict[str, Any]] = None) -> Iterator[Issue]:
         """Return an iterator of issues, optionally filtered."""
-        issues = self._issues.values()
-        
-        if filters:
-            filtered_issues = []
-            for issue in issues:
-                match = True
-                for key, value in filters.items():
-                    if key == "labels" and isinstance(value, list):
-                        # Check if any of the requested labels are in the issue's labels
-                        if not any(label in issue.labels for label in value):
-                            match = False
-                            break
-                    elif key == "status" and getattr(issue, key) != value:
-                        match = False
-                        break
-                    elif key == "assignee" and getattr(issue, key) != value:
-                        match = False
-                        break
-                if match:
-                    filtered_issues.append(issue)
-            return iter(filtered_issues)
-        
-        return iter(issues)
-    
+        raise NotImplementedError
+
     def get_issue(self, issue_id: str) -> Issue:
         """Return a specific issue by ID."""
-        if issue_id not in self._issues:
-            raise ValueError(f"Issue with ID {issue_id} not found")
-        return self._issues[issue_id]
-    
+        raise NotImplementedError
+
     def create_issue(self, title: str, description: str, **kwargs) -> Issue:
         """Create a new issue and return it."""
-        issue = MemoryIssue(title, description, self._current_user, **kwargs)
-        self._issues[issue.id] = issue
-        return issue
-    
+        raise NotImplementedError
+
     def update_issue(self, issue_id: str, **kwargs) -> Issue:
         """Update an existing issue and return the updated version."""
-        if issue_id not in self._issues:
-            raise ValueError(f"Issue with ID {issue_id} not found")
-        
-        issue = self._issues[issue_id]
-        issue.update(**kwargs)
-        return issue
-    
+        raise NotImplementedError
+
     def add_comment(self, issue_id: str, content: str) -> Comment:
         """Add a comment to an issue and return the created comment."""
-        if issue_id not in self._issues:
-            raise ValueError(f"Issue with ID {issue_id} not found")
-        
-        issue = self._issues[issue_id]
-        comment = MemoryComment(self._current_user, content)
-        issue.add_comment(comment)
-        return comment
-    
+        raise NotImplementedError
+
     def search_issues(self, query: str) -> Iterator[Issue]:
         """Search for issues matching the query string."""
-        # Simple case-insensitive search in title and description
-        query = query.lower()
-        matching_issues = []
-        
-        for issue in self._issues.values():
-            if (query in issue.title.lower() or 
-                query in issue.description.lower()):
-                matching_issues.append(issue)
-        
-        return iter(matching_issues)
+        raise NotImplementedError
+
+def get_issue_tracker_client() -> IssueTrackerClient:
+    """Return an instance of an Issue Tracker Client."""
+    from api.src.memory_tracker import MemoryIssueTrackerClient
+    return MemoryIssueTrackerClient()
